+++ jacquard/utils/vcf.py
@@ -332,23 +332,41 @@ class VcfRecord(object): #pylint: disable=too-many-instance-attributes
                          ref=self.ref,
                          alt=self.alt)
 
-    def _format_field(self):
-        """Returns string representation of format field."""
-        format_field = "."
+    def _format_tag_fields(self):
+        """Returns list of format tag names."""
+        tag_names = []
         if self.sample_tag_values:
             first_sample = list(self.sample_tag_values.keys())[0]
-            tag_names = self.sample_tag_values[first_sample].keys()
-            if tag_names:
-                format_field = ":".join(tag_names)
-        return format_field
+            tag_names = list(self.sample_tag_values[first_sample].keys())
+        return tag_names
 
-    def _sample_field(self, sample):
-        """Returns string representation of sample-format values.
+    def _sample_field(self, tag_names, sample):
+        """Returns string representation of sample-format values ordered by tag_names.
+        Tag_names must be superset of sample_tag_values.
+        Missing sample_tag_values padded as '.'.
+        If tag_names empty, returns '.'.
 
         Raises:
             KeyError: if requested sample is not defined.
+            ValueError if sample_tag_values has more keys than tag_names.
         """
-        tag_values = self.sample_tag_values[sample].values()
+        sample_tag_values = self.sample_tag_values[sample]
+        missing_tag_names = set(sample_tag_values) - set(tag_names)
+        if missing_tag_names:
+            msg = ('{}:{}:{}:{}|sample format tags are not consistent: '
+                   'requested tags [{}] but sample {} has has tags [{}] '
+                   'leaving behind [{}]')\
+                 .format(self.chrom,
+                         self.pos,
+                         self.ref,
+                         self.alt,
+                         ', '.join(tag_names),
+                         sample,
+                         ', '.join(['{}={}'.format(k,v) for k,v in sample_tag_values.items()]),
+                         ', '.join(missing_tag_names)
+                         )
+            raise ValueError(msg)
+        tag_values = [sample_tag_values.get(t, '.') for t in tag_names]
         if tag_values:
             return ":".join(tag_values)
         else:
@@ -356,14 +374,17 @@ class VcfRecord(object): #pylint: disable=too-many-instance-attributes
 
     def text(self):
         "Returns tab-delimited, newline terminated string of VcfRecord."
-        stringifier = [self.chrom, self.pos, self.vcf_id, self.ref, self.alt,
-                       self.qual, self.filter, self.info,
-                       self._format_field()]
+        tag_names = self._format_tag_fields()
+        format_field = '.' if not tag_names else ':'.join(tag_names)
+
+        fields = [self.chrom, self.pos, self.vcf_id, self.ref, self.alt,
+                  self.qual, self.filter, self.info,
+                  format_field]
 
         for sample in self.sample_tag_values:
-            stringifier.append(self._sample_field(sample))
+            fields.append(self._sample_field(tag_names, sample))
 
-        return "\t".join(stringifier) + "\n"
+        return "\t".join(fields) + "\n"
 
     def _samples_match(self, new_sample_values):
         return set(new_sample_values.keys()) == \
@@ -398,6 +419,8 @@ class VcfRecord(object): #pylint: disable=too-many-instance-attributes
             self.filter = ";".join([self.filter,
                                     new_filter])
 
+    #TODO: cgates: This is not a good equals method. Please adjust this to
+    # consider the full set of instance variables.
     def __eq__(self, other):
         return isinstance(other, VcfRecord) and self._key == other._key
 
@@ -469,4 +492,3 @@ class FileReader(object):
     def __lt__(self, other):
         key = natsort.natsort_keygen()
         return key(self.file_name) < key(other.file_name)
-